Highly refactored code
Why I've made all this changes?

Well because you know, in a good C code the function who makes the malloc also must do the free, basically the function must keep track ONLY of its memory, this helps to make everything more modular and keep the code clean, in the previous commits most of the functions would do a malloc and the function who calls it should be the responsible of freeing it. No more.

The previous commit to this I replaced most of the strcat and strcpy functions with snprintf, because those functions are not completely safe and can generate a buffer overflow or undefined behavior. And even its relative functions that takes the number of bytes to copy as parameter, like strncpy and strncat, aren't completely safe (this is more related to non null-terminated strings).

I'm not the best programmer in the world so feel free to notify me if me or my code are wrong in anything.